+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Destructure index.ts into multiple files
-- Refecture code
+- Change header transition on homepage
+- Refactor code
 
 ### Removed